+++ tests/ChannelTest.php
@@ -8,7 +8,7 @@
  */
 
 require_once __DIR__ . "/config/config.php";
-require_once __DIR__ . "/../src/mphpd.php";
+require_once __DIR__ . "/../src/MphpD.php";
 
 use FloFaber\MphpD;
 use PHPUnit\Framework\TestCase;