+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.3
 
+* [BREAKING CHANGE] Renamed the main class file `mphpd.php` to `MphpD.php`.
 * [FIX] `Floats` returned from MPD are not casted into `ints` anymore (#13). `MphpD::status` may now return `float` if `$items` contain only one item.